keyfinder_bot/requirements.txt
aiogram>=3.3.0,<4
librosa>=0.10.1,<0.11
numpy>=1.26,<3
scipy>=1.11,<2
soundfile>=0.12,<0.14
audioread>=3.0,<4
python-dotenv>=1.0,<2
typing-extensions>=4.9,<5
